All  filter + Arrow + Border click
Added an all option to the select to return all of the countries. Ensured that the all option was always at the top even after sorting the list. Changed the arrow used in the select . Started working on the option to click a border country and load it's details. Currently, the url is changing but the country details is not being updated.